test(domain): add MealPlan module to architecture tests (US-000)
