+++ src/Yumney.MealPlan.Application/Commands/GenerateShoppingListResultDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.MealPlan.Application.Commands;
+
+public sealed record GenerateShoppingListResultDto(int ItemsAdded, int StaplesSkipped);
