+++ src/Yumney.MealPlan.Application/DTOs/PlannedIngredientDto.cs
@@ -9,10 +9,3 @@ public sealed record PlannedIngredientDto(
     string? Unit,
     string Source,
     int Servings);
-
-/// <summary>
-/// All ingredients needed for a weekly meal plan.
-/// </summary>
-public sealed record PlannedShoppingListDto(
-    string Week,
-    IReadOnlyList<PlannedIngredientDto> Ingredients);
