+++ src/Yumney.MealPlan.Application/DTOs/PlannedRecipeDto.cs
@@ -10,10 +10,3 @@ public sealed record PlannedRecipeDto(
     int Servings,
     string Day,
     string MealType);
-
-/// <summary>
-/// All recipes planned for a week — input for shopping list generation.
-/// </summary>
-public sealed record WeeklyPlannedRecipesDto(
-    string Week,
-    IReadOnlyList<PlannedRecipeDto> Recipes);
