+++ src/Yumney.MealPlan.Application/DTOs/WeeklyPlanDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.MealPlan.Application.DTOs;
+
+public sealed record WeeklyPlanDto(string Week, bool IsExtendedMode, IReadOnlyList<MealSlotDto> Slots);
