+++ src/Yumney.Recipes.Api/Requests/ParseIntentRequestValidator.cs
@@ -3,9 +3,9 @@ using SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
 
 namespace SmartSolutionsLab.Yumney.Recipes.Api.Requests;
 
-public sealed class ChatRequestValidator : AbstractValidator<ChatRequestDto>
+public sealed class ParseIntentRequestValidator : AbstractValidator<ParseIntentRequestDto>
 {
-    public ChatRequestValidator()
+    public ParseIntentRequestValidator()
     {
         RuleFor(x => x.Message).NotEmpty().WithMessage("Message cannot be empty.");
     }
