+++ src/Yumney.Recipes.Application/DTOs/ChatRecipeSuggestionDto.cs
@@ -1,0 +1,6 @@
+namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
+
+public sealed record ChatRecipeSuggestionDto(
+    Guid? RecipeIdentifier,
+    string Title,
+    string? Reason);
