+++ src/Yumney.Recipes.Application/DTOs/ChatRequestDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
+
+public sealed record ChatRequestDto(string Message, IReadOnlyList<ChatMessageDto> History);
