+++ src/Yumney.Recipes.Application/DTOs/ChatResponseDto.cs
@@ -1,0 +1,5 @@
+namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
+
+public sealed record ChatResponseDto(
+    string Reply,
+    IReadOnlyList<ChatRecipeSuggestionDto> Suggestions);
