+++ src/Yumney.Recipes.Application/DTOs/ParseIntentRequestDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
+
+public sealed record ParseIntentRequestDto(string Message, string? Context);
