+++ src/Yumney.Recipes.Application/DTOs/RecognizedIngredientDto.cs
@@ -1,5 +1,3 @@
 namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
 
 public sealed record RecognizedIngredientDto(string Name, double Confidence, string? Category);
-
-public sealed record RecognizedIngredientsResponseDto(IReadOnlyList<RecognizedIngredientDto> Ingredients);
