+++ src/Yumney.Recipes.Application/DTOs/RecognizedIngredientsResponseDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.Recipes.Application.DTOs;
+
+public sealed record RecognizedIngredientsResponseDto(IReadOnlyList<RecognizedIngredientDto> Ingredients);
