+++ src/Yumney.Shared.Web/HostBuilderExtensions.cs
@@ -182,13 +182,3 @@ public static class HostBuilderExtensions
         return app;
     }
 }
-
-/// <summary>Rate limit policy names.</summary>
-public static class RateLimitPolicies
-{
-    /// <summary>Strict limit for LLM recipe import (10/min).</summary>
-    public const string RecipeImport = "RecipeImport";
-
-    /// <summary>General API rate limit (60/min).</summary>
-    public const string GeneralApi = "GeneralApi";
-}
