+++ src/Yumney.Shared/Common/IRecipeIngredientProvider.cs
@@ -10,8 +10,3 @@ public interface IRecipeIngredientProvider
         Guid recipeIdentifier,
         CancellationToken cancellationToken = default);
 }
-
-/// <summary>
-/// Lightweight ingredient info for cross-module use.
-/// </summary>
-public sealed record RecipeIngredientInfo(string Name, decimal? Amount, string? Unit, int? RecipeServings);
