+++ src/Yumney.Shared/Common/QuantityRounder.cs
@@ -25,16 +25,3 @@ public static class QuantityRounder
         return new RoundedQuantity(rounded, quantity);
     }
 }
-
-/// <summary>
-/// A display quantity with the rounded value and the original exact value.
-/// </summary>
-/// <param name="DisplayQuantity">The rounded-up quantity for display.</param>
-/// <param name="ExactQuantity">The original calculated quantity.</param>
-public sealed record RoundedQuantity(decimal DisplayQuantity, decimal ExactQuantity)
-{
-    /// <summary>
-    /// Gets a value indicating whether the quantity was rounded.
-    /// </summary>
-    public bool WasRounded => DisplayQuantity != ExactQuantity;
-}
