+++ src/Yumney.Shared/Common/RecipeIngredientInfo.cs
@@ -1,0 +1,6 @@
+namespace SmartSolutionsLab.Yumney.Shared.Common;
+
+/// <summary>
+/// Lightweight ingredient info for cross-module use.
+/// </summary>
+public sealed record RecipeIngredientInfo(string Name, decimal? Amount, string? Unit, int? RecipeServings);
