+++ src/Yumney.Shopping.Application/DTOs/ItemSourceDto.cs
@@ -1,0 +1,6 @@
+namespace SmartSolutionsLab.Yumney.Shopping.Application.DTOs;
+
+public sealed record ItemSourceDto(
+    decimal Quantity,
+    string Source,
+    DateTime OccurredAt);
