+++ src/Yumney.Shopping.Application/DTOs/MergedShoppingListDto.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.Shopping.Application.DTOs;
+
+public sealed record MergedShoppingListDto(IReadOnlyList<MergedShoppingItemDto> Items);
