+++ src/Yumney.Shopping.Infrastructure/Persistence/EventStore/ShoppingItemAddedIntegrationEvent.cs
@@ -1,0 +1,7 @@
+using SmartSolutionsLab.Yumney.Shared.Events;
+using SmartSolutionsLab.Yumney.Shopping.Domain.ShoppingLedger.Events;
+
+namespace SmartSolutionsLab.Yumney.Shopping.Infrastructure.Persistence.EventStore;
+
+#pragma warning disable SA1649
+public sealed record ShoppingItemAddedIntegrationEvent(string OwnerId, ShoppingItemAdded Inner) : IntegrationEvent;
