+++ src/Yumney.Shopping.Infrastructure/Persistence/EventStore/ShoppingItemRemovedIntegrationEvent.cs
@@ -1,0 +1,6 @@
+using SmartSolutionsLab.Yumney.Shared.Events;
+using SmartSolutionsLab.Yumney.Shopping.Domain.ShoppingLedger.Events;
+
+namespace SmartSolutionsLab.Yumney.Shopping.Infrastructure.Persistence.EventStore;
+
+public sealed record ShoppingItemRemovedIntegrationEvent(string OwnerId, ShoppingItemRemoved Inner) : IntegrationEvent;
