+++ src/Yumney.Users.Application/DTOs/SuggestionsResponseDto.cs
@@ -1,0 +1,5 @@
+namespace SmartSolutionsLab.Yumney.Users.Application.DTOs;
+
+public sealed record SuggestionsResponseDto(
+    IReadOnlyList<SuggestionDto> Suggestions,
+    IReadOnlyList<string> QuickActions);
