+++ tests/Yumney.Integration.Tests/Fixtures/AspireFixture.cs
@@ -140,10 +140,3 @@ public sealed class AspireFixture : IAsyncLifetime
         await context.SaveChangesAsync();
     }
 }
-
-[CollectionDefinition(Name)]
-#pragma warning disable CA1711 // xUnit convention requires Collection suffix
-public class AspireCollection : ICollectionFixture<AspireFixture>
-{
-    public const string Name = "Aspire";
-}
