+++ tests/Yumney.Shared.Tests/CQRS/FakeQuery.cs
@@ -1,0 +1,5 @@
+using SmartSolutionsLab.Yumney.Shared.CQRS;
+
+namespace SmartSolutionsLab.Yumney.Shared.Tests.CQRS;
+
+public sealed record FakeQuery : IQuery<int>;
